src/user_interface/ui_requirements.txt
customtkinter==5.2.2
httpx==0.27.0
pillow==10.4.0